Version: 0.1.30

+++ startUI.sh
@@ -2,5 +2,8 @@ git checkout main
 git fetch origin
 git reset --hard origin/main
 yarn 
-cp .env.production .env
+if ! test -f .env ; then 
+    cp .env.production .env
+fi
+
 NEXT_PUBLIC_HTTPS=true NODE_OPTIONS='-r @newrelic/next' node server.js
